New: set service layer in view
